+++ app/src/main/java/com/example/ars/models/IndividualUserCrop.java
@@ -10,17 +10,16 @@ public class IndividualUserCrop implements Serializable {
     private String variety;
     private String description;
 
-    @SerializedName("minTemp") private Float minTemp;
-    @SerializedName("maxTemp") private Float maxTemp;
-    @SerializedName("maxWind") private Float maxWind;
+    @SerializedName("minTemp") private Short minTemp;
+    @SerializedName("maxTemp") private Short maxTemp;
+    @SerializedName("maxWind") private Short maxWind;
 
     @SerializedName("minHumidity") private Integer minHumidity;
     @SerializedName("maxHumidity") private Integer maxHumidity;
 
-    @SerializedName("neededPrecipitation") private Float neededPrecipitation;
+    @SerializedName("neededPrecipitation") private Short neededPrecipitation;
 
     @SerializedName("sowingDepth") private Integer sowingDepth;
-
     @SerializedName("daysToGermination") private Integer daysToGermination;
     @SerializedName("daysToHarvest") private Integer daysToHarvest;
 
@@ -36,7 +35,6 @@ public class IndividualUserCrop implements Serializable {
 
     public IndividualUserCrop() {}
 
-    // Геттеры и сеттеры
     public Integer getId() { return id; }
     public void setId(Integer id) { this.id = id; }
 
@@ -52,14 +50,14 @@ public class IndividualUserCrop implements Serializable {
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
 
-    public Float getMinTemp() { return minTemp; }
-    public void setMinTemp(Float minTemp) { this.minTemp = minTemp; }
+    public Short getMinTemp() { return minTemp; }
+    public void setMinTemp(Short minTemp) { this.minTemp = minTemp; }
 
-    public Float getMaxTemp() { return maxTemp; }
-    public void setMaxTemp(Float maxTemp) { this.maxTemp = maxTemp; }
+    public Short getMaxTemp() { return maxTemp; }
+    public void setMaxTemp(Short maxTemp) { this.maxTemp = maxTemp; }
 
-    public Float getMaxWind() { return maxWind; }
-    public void setMaxWind(Float maxWind) { this.maxWind = maxWind; }
+    public Short getMaxWind() { return maxWind; }
+    public void setMaxWind(Short maxWind) { this.maxWind = maxWind; }
 
     public Integer getMinHumidity() { return minHumidity; }
     public void setMinHumidity(Integer minHumidity) { this.minHumidity = minHumidity; }
@@ -67,8 +65,8 @@ public class IndividualUserCrop implements Serializable {
     public Integer getMaxHumidity() { return maxHumidity; }
     public void setMaxHumidity(Integer maxHumidity) { this.maxHumidity = maxHumidity; }
 
-    public Float getNeededPrecipitation() { return neededPrecipitation; }
-    public void setNeededPrecipitation(Float neededPrecipitation) { this.neededPrecipitation = neededPrecipitation; }
+    public Short getNeededPrecipitation() { return neededPrecipitation; }
+    public void setNeededPrecipitation(Short neededPrecipitation) { this.neededPrecipitation = neededPrecipitation; }
 
     public Integer getSowingDepth() { return sowingDepth; }
     public void setSowingDepth(Integer sowingDepth) { this.sowingDepth = sowingDepth; }
